+++ test/fragment.spec.js
@@ -1,10 +1,10 @@
 const testFixtureMaker = require('../src/testing/test-fixture-maker');
 
 describe('Document definition fragments:', () => {
-  let testFixture;
+  const testFixture = testFixtureMaker.initFromSyncFunction('build/sync-functions/test-fragment-sync-function.js');
 
-  beforeEach(() => {
-    testFixture = testFixtureMaker.initFromSyncFunction('build/sync-functions/test-fragment-sync-function.js');
+  afterEach(() => {
+    testFixture.resetTestEnvironment();
   });
 
   it('can create documents for a document type whose definition was imported with a single-quoted filename', () => {